:truck: Extract DataCardList component from Scenario.
